completed Solidity verifier; fixed l0_eval tests; refactored some contract tests
