+++ src/contracts/sol/Constants.sol
@@ -11,8 +11,20 @@ library Constants {
         21888242871839275222246405745257275088548364400416034343698204186575808495617;
 
     // Compute this value with Fr::from(128).inverse().unwrap()
-    uint256 constant domainSizeInv = 0x300385D5FB6F3CE964DFA52B147E55AC6DE38077E8C5FDB0215A31A8C8200001;
-    uint256 constant log2DomainSize = 7;
-    uint256 constant omega = 0x16E73DFDAD310991DF5CE19CE85943E01DCB5564B6F24C799D0E470CBA9D1811;
-    uint256 constant omega_n = 0x1332CB377D53B9C681AFA4DC09F66BC37E3F2F33DEB33D9B40BD245C971B2447;
+    uint256 constant DOMAIN_SIZE_INV = 0x300385D5FB6F3CE964DFA52B147E55AC6DE38077E8C5FDB0215A31A8C8200001;
+    uint256 constant LOG2_DOMAIN_SIZE = 7;
+    uint256 constant OMEGA = 0x16E73DFDAD310991DF5CE19CE85943E01DCB5564B6F24C799D0E470CBA9D1811;
+    uint256 constant OMEGA_N = 0x1332CB377D53B9C681AFA4DC09F66BC37E3F2F33DEB33D9B40BD245C971B2447;
+
+    // These values should be overwritten with new ones from a trusted setup
+
+    // srs_g1[table_size] x and y
+    uint256 constant SRS_G1_T_X = 0x1131C3F88CCF3F0931E61C5FF5E79DF785139A75C05B514E6DC1C62019D757E6;
+    uint256 constant SRS_G1_T_Y = 0x2F34FE29E8FCF73DF75BB2D0C2AEF1E1E3E7EF4E5AEC76B5A95FF15A57066527;
+
+    // srs_g2[1] x0, x1, y0, and y1
+    uint256 constant SRS_G2_1_X_0 = 0x1B410F7D34DF6B00F6650FCC94C00196AECE9D7BEA3E675EB25C78B36C31F9C6;
+    uint256 constant SRS_G2_1_X_1 = 0x23C8353EFFE062035B257B5D6AF84E40FBC5A365FBD4EB77D8D226E4115C57BF;
+    uint256 constant SRS_G2_1_Y_0 = 0x16A44FE676145123E1EABA5A162963C5F39F26E709A049508DE4FFEBCFB2798C;
+    uint256 constant SRS_G2_1_Y_1 = 0x2A0C90C34C4DE005E7A7F740F605D0B5356BDA9C8E8AC5E29A801D9FF3CE7B47;
 }
