+++ src/contracts/tests/crypto.rs
@@ -11,41 +11,21 @@ use ark_ff::Zero;
 use ark_ff::{PrimeField, UniformRand};
 use ark_std::test_rng;
 use ethers::contract::abigen;
+use ethers::types::U256;
 use tokio::test;
 use super::setup_eth_backend;
-use crate::transcript::Transcript;
 use crate::{
     bn_solidity_utils::{f_to_u256, format_g1, format_g2},
 };
 
 abigen!(
-    BN254,
-    "./src/contracts/out/BN254.sol/BN254.json",
-);
-abigen!(
-    TestTranscript,
-    "./src/contracts/out/TestTranscript.sol/TestTranscript.json",
+    TestPairing,
+    "./src/contracts/out/TestPairing.sol/TestPairing.json",
 );
 abigen!(
     TestLagrange,
     "./src/contracts/out/TestLagrange.sol/TestLagrange.json",
 );
-
-pub fn g1_affine_to_g1point(pt: &G1Affine) -> G1Point {
-    G1Point {
-        x: f_to_u256(pt.x),
-        y: f_to_u256(pt.y),
-    }
-}
-
-pub fn g2_affine_to_g2point(pt: &G2Affine) -> G2Point {
-    G2Point {
-        x_0: f_to_u256(pt.x.c1),
-        x_1: f_to_u256(pt.x.c0),
-        y_0: f_to_u256(pt.y.c1),
-        y_1: f_to_u256(pt.y.c0),
-    }
-}
 
 #[test]
 pub async fn test_u256_conversion() {
@@ -69,7 +49,7 @@ pub async fn test_pairing() {
 
     let mut rng = test_rng();
 
-    let bn254_contract = BN254::deploy(client, ()).unwrap().send().await.unwrap();
+    let contract = TestPairing::deploy(client, ()).unwrap().send().await.unwrap();
 
     // Pairing tests that: e(-a1, a2) * e(b1, b2) * e(c2, c3) == 1
 
@@ -105,57 +85,19 @@ pub async fn test_pairing() {
     // Sanity 2
     assert_eq!(res, Fq12::one());
 
-    let result: bool = bn254_contract
-        .verify_pairing_three(
-            format_g1(a1),
-            format_g2(a2),
-            format_g1(b1),
-            format_g2(b2),
-            format_g1(c1),
-            format_g2(c2),
-        )
-        .call()
-        .await
-        .unwrap();
+    let result: bool = contract.test_pairing(
+        format_g1(a1),
+        format_g2(a2),
+        format_g1(b1),
+        format_g2(b2),
+        format_g1(c1),
+        format_g2(c2),
+    )
+    .call()
+    .await
+    .unwrap();
 
     assert!(result);
-
-    drop(anvil);
-}
-
-#[tokio::test]
-pub async fn test_transcript() {
-    let eth_backend = setup_eth_backend().await;
-    let anvil = eth_backend.0;
-    let client = eth_backend.1;
-
-    let contract = TestTranscript::deploy(client, ()).unwrap().send().await.unwrap();
-    let u1 = Fr::from(100);
-    let u2 = Fr::from(200);
-    let g1 = G1Affine::prime_subgroup_generator();
-    let g2 = G2Affine::prime_subgroup_generator();
-
-    let (ch_contract_1, ch_contract_2) = contract.test_challenges(
-        f_to_u256(u1),
-        f_to_u256(u2),
-        g1_affine_to_g1point(&g1),
-        g2_affine_to_g2point(&g2),
-    ).call().await.unwrap();
-
-    let mut transcript = Transcript::new_transcript();
-
-    transcript.update_with_u256(u1);
-    transcript.update_with_g1(&g1);
-
-    let challenge_1 = transcript.get_challenge();
-
-    transcript.update_with_u256(u2);
-    transcript.update_with_g2(&g2);
-
-    let challenge_2 = transcript.get_challenge();
-
-    assert_eq!(ch_contract_1, f_to_u256(challenge_1));
-    assert_eq!(ch_contract_2, f_to_u256(challenge_2));
 
     drop(anvil);
 }
@@ -172,8 +114,8 @@ pub async fn test_compute_l0_eval_case(alpha: Fr) {
     let anvil = eth_backend.0;
     let client = eth_backend.1;
 
-    let domain_size = 1024;
-    let log2_domain_size = 10;
+    let domain_size = 128;
+    let log2_domain_size = 7;
     let domain_size_inv = Fr::from(domain_size as u64).inverse().unwrap();
     //println!("{}", domain_size_inv);
  
@@ -202,23 +144,26 @@ pub async fn test_compute_l0_eval_case(alpha: Fr) {
     }
 
     // Step 2: Compute the value 1 / (alpha - 1)
-    let mut one_div_alpha_minus_one;
+    let mut d_minus_one_inv;
     if alpha == Fr::zero() {
-        one_div_alpha_minus_one = Fr::one().neg();
+        d_minus_one_inv = Fr::one().neg();
     } else {
-        one_div_alpha_minus_one = alpha - Fr::one();
+        d_minus_one_inv = alpha - Fr::one();
     }
-    one_div_alpha_minus_one = one_div_alpha_minus_one.inverse().unwrap();
+    d_minus_one_inv = d_minus_one_inv.inverse().unwrap();
 
     // Step 3: Compute the evaluation of the Lagrange polynomial at point alpha
-    let result = (vanishing_poly_eval * domain_size_inv) * one_div_alpha_minus_one;
+    let result = (vanishing_poly_eval * domain_size_inv) * d_minus_one_inv;
 
     assert_eq!(result, l0_eval);
 
     let contract = TestLagrange::deploy(client, ()).unwrap().send().await.unwrap();
     let onchain_result = contract.test_compute_l0_eval(
         f_to_u256(alpha),
-        f_to_u256(one_div_alpha_minus_one),
+        f_to_u256(d_minus_one_inv),
+        U256::from(log2_domain_size),
+        f_to_u256(domain_size_inv)
+        
     ).call().await.unwrap();
 
     assert_eq!(onchain_result, f_to_u256(l0_eval));
