+++ src/tests/prover_and_verifier.rs
@@ -14,7 +14,6 @@ use crate::{
 };
 use crate::prover::prover::{Prover, WitnessInput};
 use crate::verifier::Verifier;
-use crate::constants::DUMMY_VALUE;
 
 #[test]
 pub fn test_prover_and_verifier() {
@@ -42,8 +41,6 @@ pub fn test_prover_and_verifier() {
         &mut rng,
     );
 
-    let dummy_value = Fr::from(DUMMY_VALUE);
-
     let mut identity_commitments: Vec<_> = (0..table_size).map(|_| Fr::rand(&mut rng)).collect();
     let index = 10;
     identity_commitments[index] = identity_commitment;
@@ -52,7 +49,7 @@ pub fn test_prover_and_verifier() {
     let (srs_g1, srs_g2) = unsafe_setup::<Bn254, StdRng>(table_size, table_size, &mut rng);
     let pk = ProvingKey::<Bn254> { srs_g1, srs_g2: srs_g2.clone() };
 
-    let precomputed = ProverPrecomputedData::index(&pk, &mimc7.cts, dummy_value, index, &c, table_size);
+    let precomputed = ProverPrecomputedData::index(&pk, &mimc7.cts, index, &c, table_size);
 
     let witness = WitnessInput {
         identity_nullifier,
